@@ -14,7 +14,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": ">=5.4",
         "recurly/recurly-client": "2.2.*"
     },
     "require-dev": {
