@@ -13,7 +13,7 @@ abstract class AbstractAuthorizationListener implements ListenerAggregateInterfa
     /**
      * @var \Zend\Stdlib\CallbackHandler[]
      */
-    protected $listeners = array();
+    protected $listeners = [];
     
     /**
      * Logger
