@@ -19,21 +19,21 @@ class AuthenticationAdapterFactoryTest extends \PHPUnit_Framework_TestCase
     public function testCreateService()
     {
         $serviceManager = new ServiceManager();
-        $serviceManager->setService('Recurly\ModuleConfig', array(
-            'notification' => array(
-                'security' => array(
-                    'authentication' => array(
-                        'auth_adapter' => array(
-                            'config' => array(
+        $serviceManager->setService('Recurly\ModuleConfig', [
+            'notification' => [
+                'security' => [
+                    'authentication' => [
+                        'auth_adapter' => [
+                            'config' => [
                                 'accept_schemes' => 'basic',
                                 'realm'          => 'MyApp Site',
-                            ),
+                            ],
                             'passwd_file'  => __DIR__ . '/_files/passwd.txt',
-                        ),
-                    ),
-                ),
-            ),
-        ));
+                        ],
+                    ],
+                ],
+            ],
+        ]);
 
         $adapter = $this->factory->createService($serviceManager);
         $this->assertInstanceOf('Zend\Authentication\Adapter\Http', $adapter);
