@@ -21,18 +21,18 @@ class IpListenerFactoryTest extends \PHPUnit_Framework_TestCase
         $logger = $this->getMockBuilder('Zend\Log\Logger')
             ->disableOriginalConstructor()
             ->getMock();
-        
+
         $serviceManager = new ServiceManager();
         $serviceManager
-            ->setService('Recurly\ModuleConfig', array(
-                'notification' => array(
-                    'security' => array(
-                        'ip_checking' => array(
-                            'white_list' => array(),
-                        ),
-                    ),
-                ),
-            ))
+            ->setService('Recurly\ModuleConfig', [
+                'notification' => [
+                    'security' => [
+                        'ip_checking' => [
+                            'white_list' => [],
+                        ],
+                    ],
+                ],
+            ])
             ->setService('Recurly\Logger', $logger);
 
         $listener = $this->factory->createService($serviceManager);
